Fix top10emitters filter
