+++ app/javascript/app/pages/national-context/historical-emissions/historical-emissions-selectors/historical-emissions-get-selectors.js
@@ -1,3 +1,6 @@
+import { createSelector } from 'reselect';
+import { TOP_10_EMMITERS, TOP_10_EMMITERS_OPTION } from 'constants/constants';
+
 const { COUNTRY_ISO } = process.env;
 
 export const getMetadata = ({ metadata }) =>
@@ -8,3 +11,12 @@ export const getEmissionsData = ({ GHGEmissions }) =>
 export const getTargetEmissionsData = ({ GHGTargetEmissions }) =>
   GHGTargetEmissions && GHGTargetEmissions.data || null;
 export const getQuery = ({ location }) => location && location.query || null;
+
+export const getTop10EmittersOption = createSelector([ getMetadata ], meta => {
+  if (!meta) return null;
+  const value = TOP_10_EMMITERS_OPTION.value.map(p => {
+    const location = meta.location.find(l => l.label === p);
+    return location && location.value;
+  }).join();
+  return { label: TOP_10_EMMITERS, value };
+});
